+++ Ionic/clothyfie/src/app/components/users/users.component.ts
@@ -1,8 +1,10 @@
-import { Component, OnInit, ViewEncapsulation  } from '@angular/core';
+import { EditFormComponent } from './edit-form/edit-form.component';
+import { Component, OnInit, ViewEncapsulation } from '@angular/core';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { UsersService } from './../../services/user/users.service';
 import { Users } from './../../models/users';
 import { Router } from '@angular/router';
+import { ToasterService, ToasterConfig } from 'angular2-toaster';
 
 @Component({
   selector: 'app-users',
@@ -16,21 +18,34 @@ export class UsersComponent implements OnInit {
   userCount: boolean;
   spinner: boolean = true;
   selectedUser: Users[];
+  staticAlertClosed: boolean = false;
+  alertMessage: String;
 
-  constructor(private modalService: NgbModal, private userServie: UsersService, private router: Router) { }
+  public config: ToasterConfig =
+    new ToasterConfig({
+      showCloseButton: true,
+      tapToDismiss: false,
+      timeout: 2000,
+      positionClass: 'toast-bottom-center'
+    });
+
+  constructor(private modalService: NgbModal,
+     private userServie: UsersService,
+      private router: Router,
+       private ts: ToasterService) { }
 
   ngOnInit() {
     this.userServie.getUsers()
-    .subscribe(user => {
-      this.spinner = false;
-      this.users = user;
-      
-      this.userCount = this.users.length > 0 ? true : false;
-    })
+      .subscribe(user => {
+        this.spinner = false;
+        this.users = user;
+
+        this.userCount = this.users.length > 0 ? true : false;
+      });
   }
 
   openVerticallyCentered(content, user: Users[]) {
-    this.modalService.open(content, { centered: true });    
+    this.modalService.open(content, { centered: true });
     this.selectedUser = user;
   }
 
@@ -39,7 +54,17 @@ export class UsersComponent implements OnInit {
   }
 
   deleteUsers(id) {
-    this.userServie.deleteUsers(id);
+    let result = confirm('Do You Want to Delete?');
+    if (result) {
+      this.userServie.deleteUsers(id);
+      this.ts.pop('success', 'Deleted Successfully');
+    } else {
+      return;
+    }    
+  }
+
+  showAlert(event) {
+    this.ts.pop('success', event);
   }
 
 }
